third_party/blink/web_tests/fast/js/dfg-to-string-valueOf-becomes-bad-expected.txt

Tests that we do ToString conversions correctly when String.prototype.valueOf is initially fine but then gets clobbered.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "hello"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS successfullyParsed is true

TEST COMPLETE
